Restructured AEG tests to import properly
