+++ tests/AEGTree/CutNode.test.ts
@@ -1,8 +1,8 @@
 import {describe, test, expect} from "vitest";
-import {CutNode} from "../src/AEG/CutNode";
-import {AtomNode} from "../src/AEG/AtomNode";
-import {Ellipse} from "../src/AEG/Ellipse";
-import {Point} from "../src/AEG/Point";
+import {CutNode} from "../../src/AEG/CutNode";
+import {AtomNode} from "../../src/AEG/AtomNode";
+import {Ellipse} from "../../src/AEG/Ellipse";
+import {Point} from "../../src/AEG/Point";
 
 const testCenter = new Point(5, 5);
 const testEllipse = new Ellipse(testCenter, 5, 5);
